moves createBookmark function in a separate folder
