+++ src/Background/background.js
@@ -2,6 +2,7 @@ import { showNotification } from "./utils/showNotification.js";
 import { getBrowserBookmarks } from "./utils/getBrowserBookmarks.js";
 import { sendRequest } from "./requests/sendRequest.js";
 import { sendGETRequest } from "./requests/sendGetRequest.js";
+import { create_bookmarks } from "./utils/createBookmarks.js";
 
 chrome.runtime.onInstalled.addListener(function (details) {
   if (details.reason === "install") {
@@ -58,36 +59,6 @@ chrome.runtime.onMessage.addListener(async function (message) {
     }
     console.debug("adding add listener back");
     chrome.bookmarks.onCreated.addListener(sendBookmarkToElysian);
-
-    async function create_bookmarks(bookmarksData) {
-      // Start creating bookmarks in the Chrome browser
-      await createBookmarksHierarchy(bookmarksData, null);
-
-      // Reattach the event listener after all bookmarks have been created
-
-    }
-
-    function createBookmarksHierarchy(bookmarks, parentId) {
-      return Promise.all(bookmarks.map(bookmark => {
-        return new Promise((resolve) => {
-          const newBookmark = {
-            parentId: parentId || '1', // '1' is the root "Bookmarks Bar" ID
-            title: bookmark.title || "Untitled",
-            url: bookmark.url || null
-          };
-
-          chrome.bookmarks.create(newBookmark, (createdBookmark) => {
-            if (bookmark.children && bookmark.children.length > 0) {
-              // Recursively create children bookmarks
-              createBookmarksHierarchy(bookmark.children, createdBookmark.id).then(resolve);
-            } else {
-              resolve();
-            }
-          });
-        });
-      }));
-    }
-
   }
 })
